pyproject.toml
[project]
name = "truthification"
version = "0.1.0"
description = "Research project investigating truthification methods for LLMs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "wandb>=0.18.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.26.0",
    "pandas>=2.0.0",
    "tqdm>=4.66.0",
    "streamlit>=1.53.0",
    "matplotlib>=3.10.8",
    "seaborn>=0.13.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.13",
]
